# docs/swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
    "/users/admins": {
      "post": {
        "operationId": "UsersController_createAdmin",
        "summary": "Create an admin account (super_admin only).",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAdminDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "users"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/users/admins/{id}": {
      "delete": {
        "operationId": "UsersController_deleteAdmin",
        "summary": "Delete an admin account (super_admin only).",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "users"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/users/guests": {
      "post": {
        "operationId": "UsersController_createGuest",
        "summary": "Create a guest user with a QR token (admin+).",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGuestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "users"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/users": {
      "get": {
        "operationId": "UsersController_findAll",
        "summary": "List all users (admin+).",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "users"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/users/{id}": {
      "get": {
        "operationId": "UsersController_findOne",
        "summary": "Get a single user by ID (admin+).",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "users"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/users/guests/{id}/button": {
      "patch": {
        "operationId": "UsersController_setButton",
        "summary": "Enable or disable the second action button on a guest welcome page (admin+).",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetButtonDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "users"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/users/guests/{id}/seat": {
      "patch": {
        "operationId": "UsersController_setSeat",
        "summary": "Set (or clear) a guest seat number (admin+).",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetSeatDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "users"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/users/guests/{id}": {
      "delete": {
        "operationId": "UsersController_deleteGuest",
        "summary": "Delete a guest account (admin+).",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "users"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/users/me/avatar": {
      "post": {
        "operationId": "UsersController_uploadMyAvatar",
        "summary": "Upload own avatar (any authenticated user).",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "users"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/users/{id}/avatar": {
      "patch": {
        "operationId": "UsersController_uploadUserAvatar",
        "summary": "Admin sets a user avatar.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "users"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "UsersController_getAvatar",
        "summary": "Serve a user's avatar image (public).",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "users"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/users/{id}/admit": {
      "post": {
        "operationId": "UsersController_admitUser",
        "summary": "Mark a guest as admitted at the event entrance (admin+).",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "users"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Wedding Photos API",
    "description": "Backend for a wedding photo app over the Google Photos Library API. Photos live in one app-created album; the API can only see media it uploaded.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "CreateAdminDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "planner@wedding.app"
          },
          "password": {
            "type": "string",
            "example": "StrongPass123!"
          },
          "name": {
            "type": "string",
            "example": "Jane Planner"
          },
          "role": {
            "type": "string",
            "enum": [
              "admin",
              "super_admin",
              "guest"
            ],
            "default": "admin"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "CreateGuestDto": {
        "type": "object",
        "properties": {}
      },
      "SetButtonDto": {
        "type": "object",
        "properties": {}
      },
      "SetSeatDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}